[12.x] extract Backoff test to separate test
